created utils class and fixed style problems
