+++ app/src/main/java/com/example/meetup/MainActivity.java
@@ -7,7 +7,10 @@ import android.util.Log;
 
 import com.parse.ParseObject;
 
+// TODO main activity in progress
 public class MainActivity extends AppCompatActivity {
+
+    private final String TAG = "MainActivity";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -17,9 +20,9 @@ public class MainActivity extends AppCompatActivity {
         firstObject.put("message","Hey ! First message from android. Parse is now connected");
         firstObject.saveInBackground(e -> {
             if (e != null){
-                Log.e("MainActivity", e.getLocalizedMessage());
+                Log.e(TAG, e.getLocalizedMessage());
             }else{
-                Log.d("MainActivity","Object saved.");
+                Log.d(TAG,"Object saved.");
             }
         });
     }
