+++ app/src/main/java/com/example/meetup/Utils.java
@@ -1,0 +1,44 @@
+package com.example.meetup;
+
+import android.content.Context;
+import android.content.DialogInterface;
+import android.content.Intent;
+
+import androidx.appcompat.app.AlertDialog;
+
+public class Utils {
+    static void showAlert(String title, String message, Context context, Class goToClass){
+        AlertDialog.Builder builder = new AlertDialog.Builder(context)
+                .setTitle(title)
+                .setMessage(message)
+                .setPositiveButton("OK", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        dialog.cancel();
+                        Intent intent = new Intent(context, goToClass);
+                        intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
+                        context.startActivity(intent);
+                    }
+                });
+        AlertDialog ok = builder.create();
+        ok.show();
+    }
+}
+//
+//    private void showAlert(String title,String message, Context context, Class goToClass){
+//        AlertDialog.Builder builder = new AlertDialog.Builder(LoginActivity.this)
+//                .setTitle(title)
+//                .setMessage(message)
+//                .setPositiveButton("OK", new DialogInterface.OnClickListener() {
+//                    @Override
+//                    public void onClick(DialogInterface dialog, int which) {
+//                        dialog.cancel();
+//                        Intent intent = new Intent(LoginActivity.this, LogoutActivity.class);
+//                        intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
+//                        startActivity(intent);
+//                    }
+//                });
+//        AlertDialog ok = builder.create();
+//        ok.show();
+//    }
+//}
